Twitter Api Update
Following the recent changes to the twitter api, I have updated the bot so that it is able to post once again.

If your app says it is suspended on twitter dev, delete it and create a new project.

If you have any issues let me know (preferably through the discord support server https://discord.gg/jHsAW2FKnj )